refactor(adoc-core): introduce hexagonal module layout
