+++ crates/adoc-core/src/application/mod.rs
@@ -1,0 +1,2 @@
+// see ADR-0009
+pub(crate) mod compile;
