+++ crates/adoc-core/src/domain/artifact.rs
@@ -1,8 +1,7 @@
 use serde::Serialize;
 
-use crate::artifact::ArtifactWriter;
-use crate::ast::PageAst;
-use crate::diagnostic::Diagnostic;
+use crate::domain::ast::PageAst;
+use crate::domain::diagnostic::Diagnostic;
 
 #[derive(Debug, Clone, PartialEq, Eq, Serialize)]
 pub struct AgentJsonDocument {
@@ -15,21 +14,6 @@ pub struct AgentJsonDocument {
 impl AgentJsonDocument {
     pub fn to_pretty_json(&self) -> Result<String, serde_json::Error> {
         serde_json::to_string_pretty(self)
-    }
-}
-
-#[derive(Debug, Default, Clone, Copy)]
-pub(crate) struct AgentJsonArtifact;
-
-impl ArtifactWriter for AgentJsonArtifact {
-    type Output = AgentJsonDocument;
-    fn build(&self, pages: &[PageAst], diagnostics: &[Diagnostic]) -> AgentJsonDocument {
-        AgentJsonDocument {
-            schema_version: "adoc.agent.v0".to_string(),
-            pages: pages.iter().map(AgentJsonPage::from).collect(),
-            objects: Vec::new(),
-            diagnostics: diagnostics.to_vec(),
-        }
     }
 }
