+++ crates/adoc-core/src/domain/knowledge_object/mod.rs
@@ -1,0 +1,1 @@
+// Aggregate family — populated by Slice 1
