+++ crates/adoc-core/src/domain/ports/mod.rs
@@ -1,0 +1,4 @@
+// see ADR-0009
+pub(crate) mod artifact_writer;
+pub(crate) mod renderer;
+pub(crate) mod source_provider;
