+++ crates/adoc-core/src/domain/ports/renderer.rs
@@ -1,6 +1,4 @@
-pub(crate) mod html;
-
-use crate::ast::PageAst;
+use crate::domain::ast::PageAst;
 
 /// Output port for compiled pages. compile_with_provider drives the renderer
 /// through this trait so that adding a new format (e.g. Markdown export, plain
@@ -8,5 +6,3 @@ use crate::ast::PageAst;
 pub(crate) trait Renderer {
     fn render(&self, pages: &[PageAst]) -> String;
 }
-
-pub(crate) use html::HtmlRenderer;
