+++ crates/adoc-core/src/domain/scan/mod.rs
