+++ crates/adoc-core/src/infrastructure/artifact/mod.rs
@@ -1,30 +1,15 @@
 pub(crate) mod agent_json;
 
-use crate::ast::PageAst;
-use crate::diagnostic::Diagnostic;
-
 pub(crate) use agent_json::AgentJsonArtifact;
-pub use agent_json::AgentJsonDocument;
-
-/// Output port for compiler artifacts that aren't a single rendered string.
-///
-/// Each adapter declares its own `Output` shape — agent JSON today, and any
-/// future structured artifact (RDF graph, Markdown index) — so adding a new
-/// format is a new adapter rather than another edit to `compile.rs` or this
-/// trait. Static dispatch is preserved per ADR-0006: callers pick the
-/// concrete adapter at the call site.
-pub(crate) trait ArtifactWriter {
-    type Output;
-    fn build(&self, pages: &[PageAst], diagnostics: &[Diagnostic]) -> Self::Output;
-}
 
 #[cfg(test)]
 mod tests {
     use std::path::PathBuf;
 
-    use super::*;
-    use crate::ast::PageAst;
-    use crate::identity::PageId;
+    use crate::domain::ast::PageAst;
+    use crate::domain::diagnostic::Diagnostic;
+    use crate::domain::identity::PageId;
+    use crate::domain::ports::artifact_writer::ArtifactWriter;
 
     /// Stub adapter declaring a non-`AgentJsonDocument` Output. Compiling and
     /// running this proves the trait is genuinely format-agnostic.
