+++ crates/adoc-core/src/infrastructure/render/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) mod html;
+
+pub(crate) use html::HtmlRenderer;
