+++ crates/adoc-core/src/infrastructure/source/mod.rs
@@ -1,0 +1,4 @@
+// see ADR-0009
+pub(crate) mod fs;
+#[cfg(test)]
+pub(crate) mod in_memory;
