+++ crates/adoc-core/src/infrastructure/validate/mod.rs
@@ -9,24 +9,12 @@
 //! streaming context (you only know a fence is unclosed once EOF is reached),
 //! so that diagnostic remains in the parser. See ADR-0007 for the decision.
 
-use crate::ast::{BlockAst, PageAst, WorkspaceAst};
-use crate::diagnostic::{Diagnostic, DiagnosticCode, SourceSpan};
-use crate::inline::InlineSegment;
-use crate::scan::raw_html::find_raw_html;
-use crate::source::SourceFile;
-
-pub(crate) trait ValidationRule {
-    fn check(&self, page: &PageAst, source: &SourceFile, sink: &mut Vec<Diagnostic>);
-}
-
-/// Validation rule that operates on the whole `WorkspaceAst` aggregate rather
-/// than a single page — for invariants that require cross-page context (page
-/// ID uniqueness, link-target resolution, hierarchy checks). Mirrors
-/// [`ValidationRule`] so adding a workspace-level rule is a new adapter, not
-/// a branch inside the orchestrator.
-pub(crate) trait WorkspaceRule {
-    fn check(&self, workspace: &WorkspaceAst, sink: &mut Vec<Diagnostic>);
-}
+use crate::domain::ast::{BlockAst, PageAst, WorkspaceAst};
+use crate::domain::diagnostic::{Diagnostic, DiagnosticCode, SourceSpan};
+use crate::domain::inline::InlineSegment;
+use crate::domain::rules::{ValidationRule, WorkspaceRule};
+use crate::domain::scan::raw_html::find_raw_html;
+use crate::domain::source::SourceFile;
 
 /// Strict-mode page rules, applied in registration order. Adding a new rule is
 /// a new adapter plus a new entry here — no edit to `validate_page`.
@@ -193,7 +181,7 @@ mod tests {
     use std::path::PathBuf;
 
     use super::*;
-    use crate::parser::parse_page;
+    use crate::infrastructure::parser::parse_page;
 
     fn validate_text(text: &str) -> Vec<Diagnostic> {
         let source = SourceFile::new_with_identity_path(
