+++ docs/adr/0009-tactical-ddd-layout.md
@@ -1,0 +1,13 @@
+# Adopt domain/application/infrastructure module layout in adoc-core
+
+AgentDoc V0.1 introduced internal hexagonal ports (ADR-0006) but kept `adoc-core`'s files in a flat `src/` layout. V0.2 lands the first **Knowledge Object** aggregate (`Claim`), and we want the DDD building blocks visible at the file-tree level before any aggregate code is written. This ADR adopts a `domain/`/`application/`/`infrastructure/` layout in `adoc-core` and pins the placement rule for every future aggregate. Aggregates and value objects live in `domain/`; the `compile_with_provider` Application Service lives in `application/`; driven-port traits live in `domain/ports/`; their adapters and the parser, renderer, artifact-writer, validation-registry implementations live in `infrastructure/`. Validation rule traits live in `domain/rules/`; the `PAGE_RULES`/`WORKSPACE_RULES` registries and the rule structs (`RawHtmlForbidden`, `UnsafeLinkForbidden`) live in `infrastructure/validate/`. The dependency direction is `domain ← application ← infrastructure`, with `lib.rs` acting as the **composition root**: the public `compile_workspace` entry point lives in `lib.rs` as a three-line wrapper that constructs `FsSourceProvider` and delegates to `application::compile::compile_with_provider`. This keeps the Application Service's only direct construction concern off `domain` (`FsSourceProvider` is never imported from `application/compile.rs`) and concentrates infrastructure wiring in a single, reviewable place.
+
+The DDD building blocks we adopt here are the small useful set: aggregate roots with constructor-enforced invariants (`try_new` returning `Result`), value objects via single-field newtypes, an Application Service as a named layer, and Ports & Adapters per ADR-0006. We deliberately do **not** adopt Repositories (no persistence layer in v0), Domain Services (every behavior currently fits on an aggregate or rule), Domain Events (no subscribers), or Bounded-Context-per-crate (deferred until the model justifies it).
+
+The layout move itself is a single horizontal commit — Slice 0 of v0.2. **This is the only horizontal-slice exception in v0.2;** all subsequent slices are vertical end-to-end. The exception is recorded once in this ADR, not repeated. File paths churn once, reviewers gain an unambiguous home for new aggregates, and a future per-layer crate split is a `git mv` away.
+
+## Addendum: scope of Slice 0 and the application-layer adapter imports
+
+Slice 0 is mechanical file moves with zero behavior change: every test in `cargo test --workspace` remains green, the public surface re-exported from `lib.rs` keeps its current names, and `tests/public_surface.rs` compiles with no edits. The composition root is intentionally narrow — it owns the construction of `FsSourceProvider` and nothing else. `application/compile.rs` still imports `crate::infrastructure::{parser::parse_page, validate::{validate_page, validate_workspace}}` because no parser or validator port traits exist in `domain/ports/` today; introducing those abstractions is a structural addition, not a file move, and is deferred until a slice's behavior justifies the trait. The `Renderer` and `ArtifactWriter` ports do exist; their traits are imported directly from `domain::ports::{renderer, artifact_writer}` and their adapter constructors (`HtmlRenderer`, `AgentJsonArtifact`) are imported separately from `infrastructure/`. The visible payoff today is the layout invariant — every aggregate has a clear home, every adapter has a clear home, and the construction concern is concentrated in `lib.rs`.
+
+This ADR preserves and refines its predecessors. ADR-0003 (two-crate workspace) is unchanged. ADR-0005 (single public API) is preserved: `compile_workspace` remains the only public entry point. ADR-0006 (internal hexagonal ports) is preserved with port traits relocated to `domain/ports/`; visibility (`pub(crate)`) and contracts are unchanged. ADR-0007 (validation as a separate pass) is preserved with rule traits relocated to `domain/rules/` and registries to `infrastructure/validate/`. ADR-0008 (test taxonomy) is preserved; inline `#[cfg(test)]` blocks travel with their host file, and `tests/` continues to cover the public boundary.
